refactor: comments and file moves
